Split Validation's examples into individual files (#531)
This will improve the recompilation time of Validation when a change is
made to one of the examples. ( 5.9 seconds without the change vs 3.8
seconds) This will be most noticeable when working with Compose
Previews.
