+++ integration-tests/validation/src/androidTest/java/com/android/designcompose/testapp/validation/InteractionTests.kt
@@ -23,6 +23,7 @@ import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTouchInput
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.android.designcompose.DesignSettings
+import com.android.designcompose.testapp.validation.examples.InteractionTest
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
