+++ integration-tests/validation/src/main/java/com/android/designcompose/testapp/validation/examples/BlendModeTest.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.android.designcompose.testapp.validation.examples
+
+import androidx.compose.runtime.Composable
+import com.android.designcompose.annotation.DesignComponent
+import com.android.designcompose.annotation.DesignDoc
+
+@DesignDoc(id = "ZqX5i5g6inv9tANIwMMXUV")
+interface BlendModeTest {
+    @DesignComponent(node = "#stage") fun MainFrame()
+}
+
+// Demonstration of blend modes applied to different objects
+@Composable
+fun BlendModeTest() {
+    BlendModeTestDoc.MainFrame()
+}
